Cleanup PlayerTile abstraction
